Add key_management resources and routes
Adds resources intended to be used by REST API routes behind
the biome-key-management feature flag.

Also adds the key management routes, implemented using the actix
framework, to the biome::rest_api::actix module. These routes
are behind the biome-key-management and rest-api-actix feature
flags.
