+++ libsplinter/src/biome/key_management/rest_resources.rs
@@ -14,286 +14,20 @@
 
 use std::sync::Arc;
 
-use crate::actix_web::{HttpRequest, HttpResponse};
-use crate::futures::{Future, IntoFuture};
-use crate::protocol;
-use crate::rest_api::{
-    get_authorization_token, into_bytes, ErrorResponse, HandlerFunction, Method,
-    ProtocolVersionRangeGuard, Resource,
-};
+use crate::actix_web::HttpRequest;
+use crate::rest_api::get_authorization_token;
 
 use super::super::rest_api::BiomeRestConfig;
 use super::super::secrets::SecretManager;
 use super::super::sessions::{validate_token, TokenValidationError};
 
-use super::{
-    store::{KeyStore, KeyStoreError},
-    Key,
-};
-
-#[derive(Deserialize)]
-struct NewKey {
-    public_key: String,
-    encrypted_private_key: String,
-    display_name: String,
-}
-
-#[derive(Deserialize)]
-struct UpdatedKey {
-    public_key: String,
-    new_display_name: String,
-}
-
-/// Defines a REST endpoint for managing keys
-pub fn make_key_management_route(
-    rest_config: Arc<BiomeRestConfig>,
-    key_store: Arc<dyn KeyStore<Key>>,
-    secret_manager: Arc<dyn SecretManager>,
-) -> Resource {
-    Resource::build("/biome/users/{user_id}/keys")
-        .add_request_guard(ProtocolVersionRangeGuard::new(
-            protocol::BIOME_KEYS_PROTOCOL_MIN,
-            protocol::BIOME_PROTOCOL_VERSION,
-        ))
-        .add_method(
-            Method::Post,
-            handle_post(
-                rest_config.clone(),
-                key_store.clone(),
-                secret_manager.clone(),
-            ),
-        )
-        .add_method(
-            Method::Get,
-            handle_get(
-                rest_config.clone(),
-                key_store.clone(),
-                secret_manager.clone(),
-            ),
-        )
-        .add_method(
-            Method::Patch,
-            handle_patch(rest_config, key_store, secret_manager),
-        )
-}
-
-fn handle_post(
-    rest_config: Arc<BiomeRestConfig>,
-    key_store: Arc<dyn KeyStore<Key>>,
-    secret_manager: Arc<dyn SecretManager>,
-) -> HandlerFunction {
-    Box::new(move |request, payload| {
-        let key_store = key_store.clone();
-        let user_id = match request.match_info().get("user_id") {
-            Some(id) => id.to_owned(),
-            None => {
-                error!("User ID is not in path request");
-                return Box::new(
-                    HttpResponse::InternalServerError()
-                        .json(ErrorResponse::internal_error())
-                        .into_future(),
-                );
-            }
-        };
-
-        match authorize_user(&request, &user_id, &secret_manager, &rest_config) {
-            AuthorizationResult::Authorized => (),
-            AuthorizationResult::Unauthorized(msg) => {
-                return Box::new(
-                    HttpResponse::Unauthorized()
-                        .json(ErrorResponse::unauthorized(&msg))
-                        .into_future(),
-                )
-            }
-            AuthorizationResult::Failed => {
-                return Box::new(
-                    HttpResponse::InternalServerError()
-                        .json(ErrorResponse::internal_error())
-                        .into_future(),
-                );
-            }
-        }
-
-        Box::new(into_bytes(payload).and_then(move |bytes| {
-            let new_key = match serde_json::from_slice::<NewKey>(&bytes) {
-                Ok(val) => val,
-                Err(err) => {
-                    debug!("Error parsing payload {}", err);
-                    return HttpResponse::BadRequest()
-                        .json(ErrorResponse::bad_request(&format!(
-                            "Failed to parse payload: {}",
-                            err
-                        )))
-                        .into_future();
-                }
-            };
-            let key = Key::new(
-                &new_key.public_key,
-                &new_key.encrypted_private_key,
-                &user_id,
-                &new_key.display_name,
-            );
-
-            match key_store.add_key(key) {
-                Ok(()) => HttpResponse::Ok()
-                    .json(json!({ "message": "Key added successfully" }))
-                    .into_future(),
-                Err(err) => {
-                    debug!("Failed to add new key to database {}", err);
-                    match err {
-                        KeyStoreError::DuplicateKeyError(msg) => HttpResponse::BadRequest()
-                            .json(ErrorResponse::bad_request(&msg))
-                            .into_future(),
-                        KeyStoreError::UserDoesNotExistError(msg) => HttpResponse::BadRequest()
-                            .json(ErrorResponse::bad_request(&msg))
-                            .into_future(),
-                        _ => HttpResponse::InternalServerError()
-                            .json(ErrorResponse::internal_error())
-                            .into_future(),
-                    }
-                }
-            }
-        }))
-    })
-}
-
-fn handle_get(
-    rest_config: Arc<BiomeRestConfig>,
-    key_store: Arc<dyn KeyStore<Key>>,
-    secret_manager: Arc<dyn SecretManager>,
-) -> HandlerFunction {
-    Box::new(move |request, _| {
-        let key_store = key_store.clone();
-        let user_id = match request.match_info().get("user_id") {
-            Some(id) => id.to_owned(),
-            None => {
-                error!("User ID is not in path request");
-                return Box::new(
-                    HttpResponse::InternalServerError()
-                        .json(ErrorResponse::internal_error())
-                        .into_future(),
-                );
-            }
-        };
-
-        match authorize_user(&request, &user_id, &secret_manager, &rest_config) {
-            AuthorizationResult::Authorized => (),
-            AuthorizationResult::Unauthorized(msg) => {
-                return Box::new(
-                    HttpResponse::Unauthorized()
-                        .json(ErrorResponse::unauthorized(&msg))
-                        .into_future(),
-                )
-            }
-            AuthorizationResult::Failed => {
-                return Box::new(
-                    HttpResponse::InternalServerError()
-                        .json(ErrorResponse::internal_error())
-                        .into_future(),
-                );
-            }
-        }
-
-        match key_store.list_keys(Some(&user_id)) {
-            Ok(keys) => Box::new(
-                HttpResponse::Ok()
-                    .json(json!({ "data": keys }))
-                    .into_future(),
-            ),
-            Err(err) => {
-                debug!("Failed to fetch keys {}", err);
-                Box::new(
-                    HttpResponse::InternalServerError()
-                        .json(ErrorResponse::internal_error())
-                        .into_future(),
-                )
-            }
-        }
-    })
-}
-
-fn handle_patch(
-    rest_config: Arc<BiomeRestConfig>,
-    key_store: Arc<dyn KeyStore<Key>>,
-    secret_manager: Arc<dyn SecretManager>,
-) -> HandlerFunction {
-    Box::new(move |request, payload| {
-        let key_store = key_store.clone();
-        let user_id = match request.match_info().get("user_id") {
-            Some(id) => id.to_owned(),
-            None => {
-                error!("User ID is not in path request");
-                return Box::new(
-                    HttpResponse::InternalServerError()
-                        .json(ErrorResponse::internal_error())
-                        .into_future(),
-                );
-            }
-        };
-
-        match authorize_user(&request, &user_id, &secret_manager, &rest_config) {
-            AuthorizationResult::Authorized => (),
-            AuthorizationResult::Unauthorized(msg) => {
-                return Box::new(
-                    HttpResponse::Unauthorized()
-                        .json(ErrorResponse::unauthorized(&msg))
-                        .into_future(),
-                )
-            }
-            AuthorizationResult::Failed => {
-                return Box::new(
-                    HttpResponse::InternalServerError()
-                        .json(ErrorResponse::internal_error())
-                        .into_future(),
-                );
-            }
-        }
-
-        Box::new(into_bytes(payload).and_then(move |bytes| {
-            let updated_key = match serde_json::from_slice::<UpdatedKey>(&bytes) {
-                Ok(val) => val,
-                Err(err) => {
-                    debug!("Error parsing payload {}", err);
-                    return HttpResponse::BadRequest()
-                        .json(ErrorResponse::bad_request(&format!(
-                            "Failed to parse payload: {}",
-                            err
-                        )))
-                        .into_future();
-                }
-            };
-
-            match key_store.update_key(
-                &updated_key.public_key,
-                &user_id,
-                &updated_key.new_display_name,
-            ) {
-                Ok(()) => HttpResponse::Ok()
-                    .json(json!({ "message": "Key updated successfully" }))
-                    .into_future(),
-                Err(err) => {
-                    debug!("Failed to update key {}", err);
-                    match err {
-                        KeyStoreError::NotFoundError(msg) => HttpResponse::NotFound()
-                            .json(ErrorResponse::not_found(&msg))
-                            .into_future(),
-                        _ => HttpResponse::InternalServerError()
-                            .json(ErrorResponse::internal_error())
-                            .into_future(),
-                    }
-                }
-            }
-        }))
-    })
-}
-
-enum AuthorizationResult {
+pub(crate) enum AuthorizationResult {
     Authorized,
     Unauthorized(String),
     Failed,
 }
 
-fn authorize_user(
+pub(crate) fn authorize_user(
     request: &HttpRequest,
     user_id: &str,
     secret_manager: &Arc<dyn SecretManager>,
