+++ libsplinter/src/biome/rest_api/resources/mod.rs
@@ -11,3 +11,6 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
+
+#[cfg(feature = "biome-key-management")]
+pub(in super::super) mod key_management;
